requirements.txt
Flask==3.1.0
pymongo==4.10.1
Flask-Session==0.8.0
python-dotenv==1.0.1
qrcode[pil]==8.0
Pillow==11.1.0
bcrypt==4.2.1
python-dateutil==2.9.0
reportlab==4.2.5
gunicorn==23.0.0
